requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
contextual-client==0.7.0
decorator==5.2.1
distro==1.9.0
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
mypy==1.16.0
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pypdf==5.6.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.11.12
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
